Backend prototype
Adding the backend prototype, this will take a comma-separated list from the user and compare it to a couple of known skincare ingredients. If the ingredients are in the list, it will spit back out the purpose of the ingredient in the order it was found in the user inputted list.