Rehandling of Ingredients
Had to change how Ingredients and their Descriptions are handled because TableView has no easy way to just throw ArrayLists into Rows. :^)

Changed the backend prototype to work flawlessly, and now on the GUI prototype the Process button actually takes user input and handles it like so:
1. Take the new name field and add that to the Product's name variable.
2. Takes in the comma separated string and tosses it to the Product importIngList function... which converts it into an AraryList then copies it into Product's ingList variable... but only the name of the Ingredient and not the Description.

The only thing left to do is look into the database stuff, so I can compare what the user gets to what the database holds.