Finishing up GUI/backend work
Finished the GUI to be the now bare minimum to work in the context of this project! All that's left is to hook up the database/Ingredient checking functionality.

The Proceed button now copies all the information into the objects and then into the Table, and the Return button now deletes all Product information and TableView information. This allows the program to be used back and forth basically infinitely.

Thank goodness this stuff actually worked, I was a bit concerned for a hot minute.