Added database functionality
The way I'm going about this is by literally copying the database onto an ArrayList and then comparing the user's list with the "master" list that is created on application launch. Also added what ingredients my database currently is, though I will likely be expanding upon it by the time this is due.